models/alignn/requirements.txt
-e git+https://github.com/usnistgov/alignn@736ec739dfd697d64b1c2a01dc84678a24bcfacd#egg=alignn
appdirs==1.4.4
ase==3.22.1
asttokens==2.2.1
backcall==0.2.0
click==8.1.3
cmake==3.26.3
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
dgl-cu111==0.6.1
docker-pycreds==0.4.0
emmet-core==0.55.2
executing==1.2.0
filelock==3.12.0
flake8==6.0.0
fonttools==4.39.4
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.1
ipython==8.13.2
jarvis-tools==2023.4.6
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kaleido==0.2.1
kiwisolver==1.4.4
latexcodec==2.0.1
lit==16.0.5
MarkupSafe==2.1.2
-e git+ssh://git@github.com/janosh/matbench-discovery.git@e550415f6117df1bf585b5479e4cf46d65a5173d#egg=matbench_discovery
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
monty==2023.5.8
mp-api==0.33.3
mpmath==1.3.0
msgpack==1.0.5
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
palettable==3.3.3
pandas==2.0.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
plotly==5.14.1
prompt-toolkit==3.0.38
protobuf==4.23.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pybtex==0.24.0
pycodestyle==2.10.0
pydantic==1.8.1
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.15.1
pymatgen==2023.5.10
pymatviz==0.6.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytorch-ignite==0.5.0.dev20221024
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
requests==2.31.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.24.0
setproctitle==1.3.2
smmap==5.0.0
snowballstemmer==2.2.0
spglib==2.0.2
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
toolz==0.12.0
torch==1.9.0+cu111
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.6.0
tzdata==2023.3
uncertainties==3.1.7
urllib3==1.26.16
wandb==0.15.3
wcwidth==0.2.6
xmltodict==0.13.0
zipp==3.15.0
